# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WMAX API
  description: Remote Patient Monitoring — production-grade clinical signal backend
  version: 2.0.0
servers:
- url: /
  description: Standart origin (Lokal dev :8000 yoki reverse proxy ildizi)
paths:
...
  /api/v1/ingest/readings:
    post:
      tags:
      - ingest
      summary: Batch upload of 5-minute aggregates using a per-device token.
      description: 'Accepts 5-minute physiological aggregates from wearable devices.

        Idempotent: duplicate (patient_id, ts) pairs are silently ignored.

        Triggers clinical signal evaluation pipeline on every accepted batch.'
      operationId: ingest_batch_api_v1_ingest_readings_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestBatch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/ingest/sos:
    post:
      tags:
      - ingest
      summary: Raise SOS using a per-device token.
      operationId: ingest_sos_api_v1_ingest_sos_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceSosRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Ingest Sos Api V1 Ingest Sos Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/devices:
    get:
      tags:
      - devices
      summary: List device inventory with battery health and assignment status
      operationId: list_devices_api_v1_devices_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: in_stock | assigned | active | maintenance | retired
          title: Status
        description: in_stock | assigned | active | maintenance | retired
      - name: tier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: tier1_wearos | tier2_wearos_budget | tier3_ble_band
          title: Tier
        description: tier1_wearos | tier2_wearos_budget | tier3_ble_band
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response List Devices Api V1 Devices Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - devices
      summary: Register a new device into clinic inventory
      operationId: register_device_api_v1_devices_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Register Device Api V1 Devices Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/devices/enroll:
    post:
      tags:
      - devices
      summary: Create one-time enrollment code
      operationId: enroll_device_api_v1_devices_enroll_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceEnrollRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceEnrollResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/devices/claim:
    post:
      tags:
      - devices
      summary: Exchange a one-time enrollment code for a device token
      operationId: claim_device_api_v1_devices_claim_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceClaimRequest'
        required: true
      responses:
        '200':
          description: Device credential issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceClaimResponse'
        '401':
          description: Code is invalid, expired, consumed, or locked
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/devices/{device_id}/assign:
    post:
      tags:
      - devices
      summary: Assign an in-stock device to a patient for rental monitoring
      operationId: assign_device_api_v1_devices__device_id__assign_post
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAssignRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Assign Device Api V1 Devices  Device Id  Assign Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/devices/{device_id}/return:
    post:
      tags:
      - devices
      summary: Process device return from patient, inspect battery and refund deposit
      operationId: return_device_api_v1_devices__device_id__return_post
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceReturnRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Return Device Api V1 Devices  Device Id  Return Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/devices/{device_id}/release:
    post:
      tags:
      - devices
      summary: Close active device assignment
      operationId: release_device_api_v1_devices__device_id__release_post
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceReturnRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Release Device Api V1 Devices  Device Id  Release
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/devices/{device_id}/revoke-credential:
    post:
      tags:
      - devices
      summary: Revoke current device credential
      operationId: revoke_device_credential_api_v1_devices__device_id__revoke_credential_post
      security:
      - HTTPBearer: []
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Revoke Device Credential Api V1 Devices  Device Id  Revoke
                  Credential Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DeviceAssignRequest:
      properties:
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        deposit_uzs:
          type: integer
          minimum: 0.0
          title: Deposit Uzs
          description: Refundable security deposit
          default: 500000
        rental_uzs_month:
          type: integer
          minimum: 0.0
          title: Rental Uzs Month
          description: Monthly rental fee
          default: 180000
      type: object
      required:
      - patient_id
      title: DeviceAssignRequest
    DeviceCreateRequest:
      properties:
        tenant_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Tenant Id
        serial_number:
          type: string
          title: Serial Number
          description: Serial or IMEI of smart watch / band
        model_name:
          type: string
          title: Model Name
          description: E.g. Galaxy Watch 5, Xiaomi Watch 2, Medical Band
        tier:
          type: string
          enum:
          - tier1_wearos
          - tier2_wearos_budget
          - tier3_ble_band
          title: Tier
          default: tier1_wearos
        ownership:
          type: string
          title: Ownership
          description: owned | leased | patient_owned
          default: owned
        battery_health_pct:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Battery Health Pct
          default: 100
      type: object
      required:
      - serial_number
      - model_name
      title: DeviceCreateRequest
    DeviceEnrollRequest:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        ttl_minutes:
          type: integer
          maximum: 60.0
          minimum: 1.0
          title: Ttl Minutes
          default: 10
      type: object
      required:
      - device_id
      title: DeviceEnrollRequest
    DeviceEnrollResponse:
      properties:
        device_id:
          type: string
          format: uuid
        enrollment_id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^\d{6}$'
        expires_at:
          type: string
          format: date-time
      type: object
      required:
      - device_id
      - enrollment_id
      - code
      - expires_at
      title: DeviceEnrollResponse
    DeviceClaimRequest:
      properties:
        device_id:
          type: string
          format: uuid
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^\d{6}$'
      type: object
      required:
      - device_id
      - code
      title: DeviceClaimRequest
    DeviceClaimResponse:
      properties:
        device_id:
          type: string
          format: uuid
        device_token:
          type: string
      type: object
      required:
      - device_id
      - device_token
      title: DeviceClaimResponse
    DeviceReturnRequest:
      properties:
        battery_health_pct:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Battery Health Pct
        return_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Return Notes
        refund_deposit:
          type: boolean
          title: Refund Deposit
          default: true
      type: object
      title: DeviceReturnRequest
    DeviceSosRequest:
      properties:
        source:
          type: string
          title: Source
          default: watch_button
        device_lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Device Lat
        device_lon:
          anyOf:
          - type: number
          - type: 'null'
          title: Device Lon
        device_accuracy_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Device Accuracy M
      type: object
      title: DeviceSosRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IngestAck:
      properties:
        window_start:
          type: string
          format: date-time
          title: Window Start
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - window_start
      title: IngestAck
    IngestBatch:
      properties:
        batch_id:
          type: string
          format: uuid
          title: Batch Id
        sequence:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Sequence
          description: Device-local monotonic batch sequence
        patient_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Patient Id
        device_clock_utc:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Device Clock Utc
        readings:
          items:
            $ref: '#/components/schemas/ReadingIn'
          type: array
          maxItems: 500
          title: Readings
      type: object
      title: IngestBatch
    IngestResult:
      properties:
        server_time:
          type: string
          format: date-time
          title: Server Time
        batch_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Batch Id
        sequence:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sequence
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        accepted:
          items:
            type: string
            format: date-time
          type: array
          title: Accepted
        duplicate:
          items:
            type: string
            format: date-time
          type: array
          title: Duplicate
        orphaned:
          items:
            type: string
            format: date-time
          type: array
          title: Orphaned
        rejected:
          items:
            $ref: '#/components/schemas/IngestAck'
          type: array
          title: Rejected
      type: object
      required:
      - server_time
      title: IngestResult
...
    ReadingIn:
      properties:
        window_start:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window Start
        window_end:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Window End
        ts:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ts
        hr_mean:
          anyOf:
          - type: number
            maximum: 260.0
            minimum: 20.0
          - type: 'null'
          title: Hr Mean
          description: O'rtacha puls (bpm)
        hr_min:
          anyOf:
          - type: number
            maximum: 260.0
            minimum: 20.0
          - type: 'null'
          title: Hr Min
          description: Minimal puls (bpm)
        hr_max:
          anyOf:
          - type: number
            maximum: 260.0
            minimum: 20.0
          - type: 'null'
          title: Hr Max
          description: Maksimal puls (bpm)
        rmssd:
          anyOf:
          - type: number
            maximum: 600.0
            minimum: 0.0
          - type: 'null'
          title: Rmssd
          description: HRV RMSSD (ms)
        sdnn:
          anyOf:
          - type: number
            maximum: 600.0
            minimum: 0.0
          - type: 'null'
          title: Sdnn
          description: HRV SDNN (ms)
        spo2:
          anyOf:
          - type: number
            maximum: 100.0
            minimum: 40.0
          - type: 'null'
          title: Spo2
          description: Kislorod to'yinishi (%)
        skin_temp:
          anyOf:
          - type: number
            maximum: 45.0
            minimum: 25.0
          - type: 'null'
          title: Skin Temp
          description: Teri harorati (°C)
        steps:
          anyOf:
          - type: integer
            maximum: 10000.0
            minimum: 0.0
          - type: 'null'
          title: Steps
          description: 5 daqiqalik qadamlar soni
        rr_est:
          anyOf:
          - type: number
            maximum: 65.0
            minimum: 4.0
          - type: 'null'
          title: Rr Est
          description: Nafas olish tezligi (nafas/daq)
        sleep_frag:
          anyOf:
          - type: number
            maximum: 30.0
            minimum: 0.0
          - type: 'null'
          title: Sleep Frag
          description: Uyqu uzilishi indeksi
        worn_pct:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Worn Pct
        samples_n:
          type: integer
          maximum: 300.0
          minimum: 0.0
          title: Samples N
          default: 0
        worn:
          type: boolean
          title: Worn
          description: Soat taqilganmi
          default: true
        battery:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Battery
          description: Batareya quvvati (%)
      type: object
      title: ReadingIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...